feat(info): warn when the running agent is a different ghtkn build
Both versions were already in the info output, but a stale agent is easy to read
past: the output looks healthy and tokens keep coming back, they just come from
the binary the agent was started with.

`ghtkn info` now calls that case out with a warning, pointing at the restart that
fixes it. The comparison is a plain inequality rather than an ordering: either
direction means the agent is not running this ghtkn's code, and saying which side
is older would need semantic version parsing that buys nothing here. It is skipped
unless both versions are known, so a development build (which reports "unknown")
stays quiet.

The warning goes to the logger, i.e. stderr, so the JSON on stdout is untouched
and `ghtkn info | jq` keeps working.
