+++ pkg/cli/info/command_internal_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	agentapi "github.com/suzuki-shunsuke/ghtkn-go-sdk/ghtkn/backend/agent"
+	"github.com/suzuki-shunsuke/ghtkn/pkg/agent/server"
 	"github.com/suzuki-shunsuke/ghtkn/pkg/controller/info"
 )
 
@@ -23,6 +24,68 @@ func TestFormatTTLDays(t *testing.T) {
 		if got := formatTTLDays(tt.d); got != tt.want {
 			t.Errorf("formatTTLDays(%s) = %q, want %q", tt.d, got, tt.want)
 		}
+	}
+}
+
+func TestStaleAgent(t *testing.T) {
+	t.Parallel()
+	tests := []struct {
+		name         string
+		agentVersion string
+		ghtknVersion string
+		want         bool
+	}{
+		{
+			name:         "same version",
+			agentVersion: "v0.3.4",
+			ghtknVersion: "v0.3.4",
+		},
+		{
+			name:         "older agent",
+			agentVersion: "v0.3.1",
+			ghtknVersion: "v0.3.4",
+			want:         true,
+		},
+		{
+			// Not an ordering: a newer agent is just as much "not this ghtkn's code".
+			name:         "newer agent",
+			agentVersion: "v0.4.0",
+			ghtknVersion: "v0.3.4",
+			want:         true,
+		},
+		{
+			name:         "agent too old to report a version",
+			agentVersion: "",
+			ghtknVersion: "v0.3.4",
+		},
+		{
+			name:         "agent built without version information",
+			agentVersion: server.UnknownVersion,
+			ghtknVersion: "v0.3.4",
+		},
+		{
+			name:         "ghtkn built without version information",
+			agentVersion: "v0.3.4",
+			ghtknVersion: server.UnknownVersion,
+		},
+		{
+			name:         "both built without version information",
+			agentVersion: server.UnknownVersion,
+			ghtknVersion: server.UnknownVersion,
+		},
+		{
+			name:         "no ghtkn version",
+			agentVersion: "v0.3.4",
+			ghtknVersion: "",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			if got := staleAgent(tt.agentVersion, tt.ghtknVersion); got != tt.want {
+				t.Errorf("staleAgent(%q, %q) = %v, want %v", tt.agentVersion, tt.ghtknVersion, got, tt.want)
+			}
+		})
 	}
 }
